+++ src/logger.js
@@ -1,0 +1,7 @@
+const bunyan = require('bunyan')
+const logger = bunyan.createLogger({
+  name: '@eventzimmer/aggregator',
+  level: (process.env.LOG_LEVEL !== undefined) ? process.env.LOG_LEVEL : 'debug'
+})
+
+module.exports = logger